+++ cli/cli/commands/enclave/new/new.go
@@ -3,15 +3,13 @@ package new
 import (
 	"context"
 	"fmt"
-	"github.com/docker/docker/client"
-	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager"
+	"github.com/kurtosis-tech/container-engine-lib/lib"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/defaults"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/engine_manager"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/execution_ids"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/logrus_log_levels"
 	"github.com/kurtosis-tech/kurtosis-engine-api-lib/api/golang/kurtosis_engine_rpc_api_bindings"
-	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
@@ -24,7 +22,7 @@ const (
 	isPartitioningEnabledArg = "with-partitioning"
 
 	defaultIsPartitioningEnabled = false
-	shouldPublishPorts = true
+	shouldPublishPorts           = true
 )
 
 var apiContainerVersion string
@@ -32,9 +30,9 @@ var isPartitioningEnabled bool
 var kurtosisLogLevelStr string
 
 var NewCmd = &cobra.Command{
-	Use:                   command_str_consts.EnclaveNewCmdStr,
-	Short:                 "Creates a new, empty Kurtosis enclave",
-	RunE:                  run,
+	Use:   command_str_consts.EnclaveNewCmdStr,
+	Short: "Creates a new, empty Kurtosis enclave",
+	RunE:  run,
 }
 
 func init() {
@@ -68,17 +66,12 @@ func run(cmd *cobra.Command, args []string) error {
 
 	ctx := context.Background()
 
-	dockerClient, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
+	kurtosisBackend, err := lib.GetLocalDockerKurtosisBackend()
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred creating the Docker client")
+		return stacktrace.Propagate(err, "An error occurred getting a Kurtosis backend connected to local Docker")
 	}
-	dockerManager := docker_manager.NewDockerManager(
-		logrus.StandardLogger(),
-		dockerClient,
-	)
-	engineManager := engine_manager.NewEngineManager(dockerManager)
-	objAttrsProvider := schema.GetObjectAttributesProvider()
-	engineClient, closeClientFunc, err := engineManager.StartEngineIdempotentlyWithDefaultVersion(ctx, objAttrsProvider, defaults.DefaultEngineLogLevel)
+	engineManager := engine_manager.NewEngineManager(kurtosisBackend)
+	engineClient, closeClientFunc, err := engineManager.StartEngineIdempotentlyWithDefaultVersion(ctx, defaults.DefaultEngineLogLevel)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred creating a new Kurtosis engine client")
 	}