+++ cli/cli/helpers/engine_manager/engine_manager.go
@@ -3,66 +3,52 @@ package engine_manager
 import (
 	"context"
 	"fmt"
-	"github.com/docker/go-connections/nat"
-	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/container_status"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/engine"
 	"github.com/kurtosis-tech/kurtosis-engine-api-lib/api/golang/kurtosis_engine_rpc_api_bindings"
-	"github.com/kurtosis-tech/object-attributes-schema-lib/forever_constants"
 	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/emptypb"
-	"net"
 	"strconv"
 	"strings"
 	"time"
 )
 
 const (
-
 	waitForEngineResponseTimeout = 5 * time.Second
-	shouldGetStoppedContainersWhenCheckingForExistingEngines = false
-
-	engineStopTimeout = 30 * time.Second
-
-	// Engine container port number string parsing constants
-	hostMachinePortNumStrParsingBase = 10
-	hostMachinePortNumStrParsingBits = 16
 
 	// --------------------------- Old port parsing constants ------------------------------------
 	// These are the old labels that the API container used to use before 2021-11-15 for declaring its port num protocol
 	// We can get rid of this after 2022-05-15, when we're confident no users will be running API containers with the old label
-	pre2021_11_15_portNum = uint16(9710)
+	pre2021_11_15_portNum   = uint16(9710)
 	pre2021_11_15_portProto = schema.PortProtocol_TCP
 
 	// These are the old labels that the API container used to use before 2021-12-02 for declaring its port num protocol
 	// We can get rid of this after 2022-06-02, when we're confident no users will be running API containers with the old label
-	pre2021_12_02_portNumLabel      = "com.kurtosistech.port-number"
-	pre2021_12_02_portNumBase = 10
+	pre2021_12_02_portNumLabel    = "com.kurtosistech.port-number"
+	pre2021_12_02_portNumBase     = 10
 	pre2021_12_02_portNumUintBits = 16
-	pre2021_12_02_portProtocol = schema.PortProtocol_TCP
+	pre2021_12_02_portProtocol    = schema.PortProtocol_TCP
 	// --------------------------- Old port parsing constants ------------------------------------
 )
 
 // Unfortunately, Docker doesn't have constants for the protocols it supports declared
 var objAttrsSchemaPortProtosToDockerPortProtos = map[schema.PortProtocol]string{
-	schema.PortProtocol_TCP: "tcp",
+	schema.PortProtocol_TCP:  "tcp",
 	schema.PortProtocol_SCTP: "sctp",
-	schema.PortProtcol_UDP: "udp",
-}
-
-var engineContainerLabels = map[string]string{
-	forever_constants.AppIDLabel: forever_constants.AppIDValue,
-	forever_constants.ContainerTypeLabel: forever_constants.ContainerType_EngineServer,
+	schema.PortProtcol_UDP:   "udp",
 }
 
 type EngineManager struct {
-	dockerManager *docker_manager.DockerManager
+	kurtosisBackend backend_interface.KurtosisBackend
 	// Make engine IP, port, and protocol configurable in the future
 }
 
-func NewEngineManager(dockerManager *docker_manager.DockerManager) *EngineManager {
-	return &EngineManager{dockerManager: dockerManager}
+func NewEngineManager(kurtosisBackend backend_interface.KurtosisBackend) *EngineManager {
+	return &EngineManager{kurtosisBackend: kurtosisBackend}
 }
 
 /*
@@ -70,11 +56,11 @@ Returns:
 	- The engine status
 	- The host machine port bindings (not present if the engine is stopped)
 	- The engine version (only present if the engine is running)
- */
+*/
 func (manager *EngineManager) GetEngineStatus(
 	ctx context.Context,
 ) (EngineStatus, *hostMachineIpAndPort, string, error) {
-	runningEngineContainers, err := manager.dockerManager.GetContainersByLabels(ctx, engineContainerLabels, shouldGetStoppedContainersWhenCheckingForExistingEngines)
+	runningEngineContainers, err := manager.kurtosisBackend.GetEngines(ctx, getRunningEnginesFilter())
 	if err != nil {
 		return "", nil, "", stacktrace.Propagate(err, "An error occurred getting Kurtosis engine containers")
 	}
@@ -82,69 +68,16 @@ func (manager *EngineManager) GetEngineStatus(
 	numRunningEngineContainers := len(runningEngineContainers)
 	if numRunningEngineContainers > 1 {
 		return "", nil, "", stacktrace.NewError("Cannot report engine status because we found %v running Kurtosis engine containers; this is very strange as there should never be more than one", numRunningEngineContainers)
-	}
-	if numRunningEngineContainers == 0 {
+	} else if numRunningEngineContainers == 0 {
 		return EngineStatus_Stopped, nil, "", nil
 	}
-	engineContainer := runningEngineContainers[0]
-
-	currentlyRunningEngineContainerLabels := engineContainer.GetLabels()
-	objAttrPrivatePort, err := getPrivateEnginePort(currentlyRunningEngineContainerLabels)
-	if err != nil {
-		return "", nil, "", stacktrace.Propagate(err, "An error occurred getting the engine container's object attributes schema private port")
-	}
-
-	privatePortObjAttrProto := objAttrPrivatePort.GetProtocol()
-	privatePortDockerProto, found := objAttrsSchemaPortProtosToDockerPortProtos[privatePortObjAttrProto]
-	if !found {
-		return "", nil, "", stacktrace.NewError(
-			"No Docker protocol was defined for obj attr proto '%v'; this is a bug in Kurtosis",
-			privatePortObjAttrProto,
-		)
-	}
-
-	privatePortNumStr := fmt.Sprintf("%v", objAttrPrivatePort.GetNumber())
-	dockerPrivatePort, err := nat.NewPort(
-		privatePortDockerProto,
-		privatePortNumStr,
-	)
-	if err != nil {
-		return "", nil, "", stacktrace.Propagate(
-			err,
-			"An error occurred creating the engine container Docker private port object from port number '%v' and protocol '%v', which is necessary for getting its host machine port bindings",
-			privatePortNumStr,
-			privatePortDockerProto,
-		)
-	}
-
-	hostMachineEnginePortBinding, found := engineContainer.GetHostPortBindings()[dockerPrivatePort]
-	if !found {
-		return "", nil, "", stacktrace.NewError("Found a Kurtosis engine server container, but it didn't have a host machine port binding - this is likely a Kurtosis bug")
-	}
-
-	hostMachineIpAddrStr := hostMachineEnginePortBinding.HostIP
-	hostMachineIp := net.ParseIP(hostMachineIpAddrStr)
-	if hostMachineIp == nil {
-		return "", nil, "", stacktrace.NewError("We got host machine IP '%v' for accessing the engine container, but it wasn't a valid IP", hostMachineIpAddrStr)
-	}
-
-	hostMachinePortNumStr := hostMachineEnginePortBinding.HostPort
-	hostMachinePortNumUint64, err := strconv.ParseUint(hostMachinePortNumStr, hostMachinePortNumStrParsingBase, hostMachinePortNumStrParsingBits)
-	if err != nil {
-		return "", nil, "", stacktrace.Propagate(
-			err,
-			"An error occurred parsing engine container host machine port num string '%v' using base '%v' and num bits '%v'",
-			hostMachinePortNumStr,
-			hostMachinePortNumStrParsingBase,
-			hostMachinePortNumStrParsingBits,
-		)
-	}
-	hostMachinePortNumUint16 := uint16(hostMachinePortNumUint64) // Okay to do due to specifying the number of bits above
+	engineContainer := getFirstEngineFromMap(runningEngineContainers)
 
 	runningEngineIpAndPort := &hostMachineIpAndPort{
-		ipAddr:  hostMachineIp,
-		portNum: hostMachinePortNumUint16,
+		ipAddr:  engineContainer.GetPublicIPAddress(),
+		portNum: engineContainer.GetPublicGRPCPort().GetNumber(),
 	}
+
 	engineClient, clientCloseFunc, err := getEngineClientFromHostMachineIpAndPort(runningEngineIpAndPort)
 	if err != nil {
 		return EngineStatus_ContainerRunningButServerNotResponding, runningEngineIpAndPort, "", nil
@@ -160,7 +93,7 @@ func (manager *EngineManager) GetEngineStatus(
 }
 
 // Starts an engine if one doesn't exist already, and returns a client to it
-func (manager *EngineManager) StartEngineIdempotentlyWithDefaultVersion(ctx context.Context, objAttrsProvider schema.ObjectAttributesProvider, logLevel logrus.Level) (kurtosis_engine_rpc_api_bindings.EngineServiceClient, func() error, error) {
+func (manager *EngineManager) StartEngineIdempotentlyWithDefaultVersion(ctx context.Context, logLevel logrus.Level) (kurtosis_engine_rpc_api_bindings.EngineServiceClient, func() error, error) {
 	status, maybeHostMachinePortBinding, engineVersion, err := manager.GetEngineStatus(ctx)
 	if err != nil {
 		return nil, nil, stacktrace.Propagate(err, "An error occurred retrieving the Kurtosis engine status, which is necessary for creating a connection to the engine")
@@ -168,8 +101,7 @@ func (manager *EngineManager) StartEngineIdempotentlyWithDefaultVersion(ctx cont
 	engineGuarantor := newEngineExistenceGuarantorWithDefaultVersion(
 		ctx,
 		maybeHostMachinePortBinding,
-		manager.dockerManager,
-		objAttrsProvider,
+		manager.kurtosisBackend,
 		logLevel,
 		engineVersion,
 	)
@@ -181,7 +113,7 @@ func (manager *EngineManager) StartEngineIdempotentlyWithDefaultVersion(ctx cont
 }
 
 // Starts an engine if one doesn't exist already, and returns a client to it
-func (manager *EngineManager) StartEngineIdempotentlyWithCustomVersion(ctx context.Context, objAttrsProvider schema.ObjectAttributesProvider, engineImageVersionTag string, logLevel logrus.Level) (kurtosis_engine_rpc_api_bindings.EngineServiceClient, func() error, error) {
+func (manager *EngineManager) StartEngineIdempotentlyWithCustomVersion(ctx context.Context, engineImageVersionTag string, logLevel logrus.Level) (kurtosis_engine_rpc_api_bindings.EngineServiceClient, func() error, error) {
 	status, maybeHostMachinePortBinding, engineVersion, err := manager.GetEngineStatus(ctx)
 	if err != nil {
 		return nil, nil, stacktrace.Propagate(err, "An error occurred retrieving the Kurtosis engine status, which is necessary for creating a connection to the engine")
@@ -189,8 +121,7 @@ func (manager *EngineManager) StartEngineIdempotentlyWithCustomVersion(ctx conte
 	engineGuarantor := newEngineExistenceGuarantorWithCustomVersion(
 		ctx,
 		maybeHostMachinePortBinding,
-		manager.dockerManager,
-		objAttrsProvider,
+		manager.kurtosisBackend,
 		engineImageVersionTag,
 		logLevel,
 		engineVersion,
@@ -202,39 +133,19 @@ func (manager *EngineManager) StartEngineIdempotentlyWithCustomVersion(ctx conte
 	return engineClient, engineClientCloseFunc, nil
 }
 
-
 // Stops the engine if it's running, doing nothing if not
 func (manager *EngineManager) StopEngineIdempotently(ctx context.Context) error {
-	matchingEngineContainers, err := manager.dockerManager.GetContainersByLabels(
-		ctx,
-		engineContainerLabels,
-		shouldGetStoppedContainersWhenCheckingForExistingEngines,
-	)
+	_, erroredEngineIds, err := manager.kurtosisBackend.StopEngines(ctx, getRunningEnginesFilter())
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting Kurtosis engine containers, which we need to check for existing engines")
+		return stacktrace.Propagate(err, "An error occurred stopping ")
 	}
-
-	numMatchingEngineContainers := len(matchingEngineContainers)
-	if numMatchingEngineContainers == 0 {
-		return nil
-	}
-	if numMatchingEngineContainers > 1 {
-		logrus.Warnf(
-			"Found %v Kurtosis engine containers, which is strange because there should never be more than 1 engine container; all will be stopped",
-			numMatchingEngineContainers,
-		)
-	}
-
 	engineStopErrorStrs := []string{}
-	for _, engineContainer := range matchingEngineContainers {
-		containerName := engineContainer.GetName()
-		containerId := engineContainer.GetId()
-		if err := manager.dockerManager.StopContainer(ctx, containerId, engineStopTimeout); err != nil {
+	for engineId, err := range erroredEngineIds {
+		if err != nil {
 			wrappedErr := stacktrace.Propagate(
 				err,
-				"An error occurred stopping engine container '%v' with ID '%v'",
-				containerName,
-				containerId,
+				"An error occurred stopping engine container `%v'",
+				engineId,
 			)
 			engineStopErrorStrs = append(engineStopErrorStrs, wrappedErr.Error())
 		}
@@ -249,6 +160,7 @@ func (manager *EngineManager) StopEngineIdempotently(ctx context.Context) error 
 			),
 		)
 	}
+
 	return nil
 }
 
@@ -331,7 +243,6 @@ func getPrivateEnginePort(containerLabels map[string]string) (*schema.PortSpec, 
 	return pre2021_11_15Port, nil
 }
 
-
 func getApiContainerPrivatePortUsingPre2021_12_02Label(containerLabels map[string]string) (*schema.PortSpec, error) {
 	// We can get rid of this after 2022-06-02, when we're confident no users will be running API containers with this label
 	portNumStr, found := containerLabels[pre2021_12_02_portNumLabel]
@@ -342,7 +253,7 @@ func getApiContainerPrivatePortUsingPre2021_12_02Label(containerLabels map[strin
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred parsing pre-2021-12-02 private port num string '%v' to a uint16", portNumStr)
 	}
-	portNumUint16 := uint16(portNumUint64)  // Safe to do because we pass in the number of bits to the ParseUint call above
+	portNumUint16 := uint16(portNumUint64) // Safe to do because we pass in the number of bits to the ParseUint call above
 	result, err := schema.NewPortSpec(portNumUint16, pre2021_12_02_portProtocol)
 	if err != nil {
 		return nil, stacktrace.Propagate(
@@ -353,4 +264,24 @@ func getApiContainerPrivatePortUsingPre2021_12_02Label(containerLabels map[strin
 		)
 	}
 	return result, nil
+}
+
+// getRunningEnginesFilter returns a filter for engines with status engine.EngineStatus_Running
+func getRunningEnginesFilter() *engine.EngineFilters {
+	return &engine.EngineFilters{
+		Statuses: map[container_status.ContainerStatus]bool{
+			container_status.ContainerStatus_Running: true,
+		},
+	}
+}
+
+// getFirstEngineFromMap returns the first value iterated by the `range` statement on a map
+// returns nil if the map is empty
+func getFirstEngineFromMap(engineMap map[string]*engine.Engine) *engine.Engine {
+	firstEngineInMap := (*engine.Engine)(nil)
+	for _, engineInMap := range engineMap {
+		firstEngineInMap = engineInMap
+		break
+	}
+	return firstEngineInMap
 }